// src/interfaces/ILibros.ts
import { Categoria } from '../../../api-libreria/src/modulos/categorias/entities/categoria.entity';
import { ICategoria } from './ICategoria';


// Generated by https://quicktype.io


export interface ILibro {
    isbn:             string;
    title:            string;
    pageCount:        number;
    precio:           string;
    publishedDate:    string;
    thumbnailUrl:     string;
    shortDescription: string;
    longDescription:  string;
    status:           string;
    key?:             string;
    categoria:        ICategoria;
}
export interface IColumn {
    key: string;
    label: string;
}
const ColumLibro:IColumn[] = [
    {
      key: "isbn",
      label: "ISBN",
    },
    {
      key: "title",
      label: "TITLE",
    },
    {
      key: "page",
      label: "PAGE",
    },
    {
        key: "precio",
        label: "PRECIO",
      },
    {
        key: "thumbnailUrl",
        label: "Imagen",
    },
  ];

  export default ColumLibro;